backend/requirements.txt
Flask
Flask-SocketIO
python-dotenv
python-engineio
python-socketio
gevent
gevent-websocket
litellm
pytest